Refactor: 게시판 조회 코드 리뷰 반영
