+++ src/interfaces/board.ts
